Address last comments by JM
