+++ forte/solvers/solver.py
@@ -190,32 +190,10 @@ class Solver(ABC):
         This function converts the input of a user into the standard
         input for a multi-state computation.
 
-        Multi-state computations need a dict[StateInfo,list(float)] object,
-        which can be cumbersome to write out. This function accepts various inputs,
-        for example:
-            >>> # single state (one eigenstate of state ``state``)
-            >>> state = StateInfo(...)
-            >>> _parse_states(state)
-
-            >>> # list of states (one eigenstate of ``state_1`` and one of ``state_2``)
-            >>> state_1 = StateInfo(...)
-            >>> state_2 = StateInfo(...)
-            >>> _parse_states([state_1,state_2])
-
-            >>> # dict of states (5 eigenstate of ``state_1`` and 3 of ``state_2``)
-            >>> state_info_1 = StateInfo(...)
-            >>> state_info_2 = StateInfo(...)
-            >>> _parse_states({state_info_1: 5,state_info_2: 3})
-
-            >>> # dict of states (5 eigenstate of ``state_1`` and 3 of ``state_2``)
-            >>> state_info_1 = StateInfo(...)
-            >>> state_info_2 = StateInfo(...)
-            >>> _parse_states({state_info_1: [1.0,1.0,0.5,0.5,0.5],state_info_2: [0.25,0.25,0.25]})
-
         Parameters
         ----------
-        states: (StateInfo, list(StateInfo), dict[StateInfo,int], dict[StateInfo,list(float)]
-            The user input can be one of three options:
+        states: StateInfo, or list(StateInfo), or dict[StateInfo,int], or dict[StateInfo,list(float)]
+            The user input can be one of four options:
             1. A single state
             2. A list of single states (will compute one level for each type of state)
             3. A dictionary that maps StateInfo objects to the number of states to compute
